cleanup(change_detection): moved pipes-related tests to the pipes dir
