+++ modules/angular2/test/change_detection/pipes/pipe_registry_spec.js
